ENH: added functions to allow discovery of supported methods
