+++ skbio/diversity/__init__.py
@@ -24,6 +24,8 @@ Functions
 
     alpha_diversity
     beta_diversity
+    get_alpha_diversity_metrics
+    get_beta_diversity_metrics
 
 Examples
 --------
@@ -255,8 +257,10 @@ from __future__ import absolute_import, division, print_function
 
 from skbio.util import TestRunner
 
-from ._driver import alpha_diversity, beta_diversity
+from ._driver import (alpha_diversity, beta_diversity,
+                      get_alpha_diversity_metrics, get_beta_diversity_metrics)
 
-__all__ = ["alpha_diversity", "beta_diversity"]
+__all__ = ["alpha_diversity", "beta_diversity", "get_alpha_diversity_metrics",
+           "get_beta_diversity_metrics"]
 
 test = TestRunner(__file__).test
