+++ skbio/diversity/alpha/_gini.py
@@ -10,7 +10,7 @@ from __future__ import absolute_import, division, print_function
 
 import numpy as np
 
-from skbio.diversity.alpha._base import _validate_counts_vector
+from skbio.diversity._validate import _validate_counts_vector
 from skbio.util._decorator import experimental
